World Size Type Change

+++ include/world.hpp
@@ -13,7 +13,7 @@ class Game;
 
 class World {
 public:
-    static constexpr uint16_t WIDTH = 512, HEIGHT = 24;
+    static constexpr int WIDTH = 512, HEIGHT = 24;
 
     // Returns a texture based on the given type, like a map.
     Texture2D* blockTextures[COUNT];
